+++ src/deion/worker/util/defaultGameAttributes.ts
@@ -95,6 +95,7 @@ const defaultGameAttributes: GameAttributesLeague = {
 	threePointTendencyFactor: 1,
 	threePointAccuracyFactor: 1,
 	twoPointAccuracyFactor: 1,
+	triggeredEvents: [],
 
 	// These will always be overwritten when creating a league, just here for TypeScript
 	lid: 0,